feat(profile menu): change for fsd
