+++ src/app/router/PublicRoutes.tsx
@@ -1,0 +1,10 @@
+import { Navigate, Outlet, useLocation } from 'react-router-dom';
+
+const PublicRoutes = () => {
+  const location = useLocation();
+  const token = localStorage.getItem('accessToken');
+
+  return <>{token ? <Navigate to="/home" state={{ from: location }} replace /> : <Outlet />}</>;
+};
+
+export default PublicRoutes;
