--- src/components/profile/Profile.module.scss
@@ -1,5 +1,0 @@
-.profile {
-    position: absolute;
-    right: 39px;
-    display: flex;
-}
