--- src/components/profile/Profile.tsx
@@ -1,12 +1,0 @@
-import ProfileMenu from '../navigates/profile/ProfileMenu';
-import style from './Profile.module.scss'
-
-const Profile = () => {
-    return (
-        <div className={style.profile}>
-           <ProfileMenu />
-        </div>
-    );
-};
-
-export default Profile;
