+++ src/components/ui/buttons/Buttons.module.scss
@@ -52,7 +52,7 @@
   background-color: transparent;
   font-size: 14px;
   &:hover {
-    color: #3cb3a1;
+    color: var(--orange-500);
   }
 }
