--- src/routes/PublicRoutes.tsx
@@ -1,17 +1,0 @@
-import { Outlet, useLocation, Navigate } from 'react-router-dom';
-
-const PublicRoutes = () => {
-    const location = useLocation();
-    const token = localStorage.getItem('accessToken');
-
-    return (
-        <>
-        {token
-        ? <Navigate to="/home" state={{ from: location }} replace />
-        : <Outlet />
-        }
-        </>
-    )
-}
-
-export default PublicRoutes
