+++ src/widgets/app-layout/ui/Layout.module.scss
