+++ src/widgets/header/ui/Profile/Profile.module.scss
@@ -1,0 +1,80 @@
+.profile {
+  position: absolute;
+  right: 39px;
+  display: flex;
+  .content {
+    position: relative;
+    .icon {
+      color: var(--gray-800);
+      font-size: 30px;
+      padding-right: 5px;
+      border: none;
+      background-color: transparent;
+      transition: all 0.2s ease;
+      cursor: pointer;
+      &:hover {
+        color: #686476;
+      }
+    }
+    .menu {
+      position: absolute;
+      width: 200px;
+      top: 36px;
+      right: 0;
+      padding: 0 0 15px;
+      border-radius: 12px;
+      background-color: #ffffff;
+      overflow: hidden;
+      box-shadow:
+        rgba(14, 63, 126, 0.04) 0px 0px 0px 1px,
+        rgba(42, 51, 69, 0.04) 0px 2px 2px -1px,
+        rgba(42, 51, 70, 0.04) 0px 10px 10px -4px,
+        rgba(42, 51, 70, 0.04) 0px 20px 20px -8px,
+        rgba(14, 63, 126, 0.04) 0px 30px 30px -12px,
+        rgba(14, 63, 126, 0.04) 0px 50px 50px -20px;
+      z-index: 99;
+      .username {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        font-size: 15px;
+        font-weight: 600;
+        color: #fff;
+        text-align: center;
+        padding: 18px 6px;
+        background-color: var(--gray-600);
+      }
+      li {
+        padding: 17px 12px;
+        a {
+          display: flex;
+          align-items: center;
+          color: #363b37;
+          font-size: 12.5px;
+          text-decoration: none;
+          .icon {
+            font-size: 20px;
+            color: inherit;
+            transition: color 0.2s ease;
+          }
+          &:hover {
+            color: var(--orange-500);
+          }
+        }
+        &:hover {
+          background-color: #fafbff;
+        }
+      }
+      li:not(:last-child) {
+        border-bottom: 1px solid #f1f3f7;
+      }
+      .logout {
+        display: flex;
+        justify-content: center;
+      }
+    }
+    .hidden {
+      display: none;
+    }
+  }
+}
