+++ src/widgets/header/ui/Search/Search.module.scss
